requirements.txt
autopep8==2.0.0
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.9.1
pymongo==4.3.2
python-dotenv==0.21.0
tomli==2.0.1
Werkzeug==2.2.2
